added safe teleport
